+++ R/html_document_base.R
@@ -62,13 +62,13 @@ html_document_base <- function(theme = NULL,
   old_theme <- NULL
   pre_knit <- function(input, ...) {
     if (is_bs_themeish(theme)) {
-      old_theme <<- bootstraplib::bs_global_set(as_bs_theme(theme))
+      old_theme <<- bslib::bs_global_set(as_bs_theme(theme))
     }
   }
   post_knit <- function(metadata, input_file, runtime, ...) {}
   on_exit <- function() {
     # In this case, we know we've altered global state, so restore the old theme
-    if (is_bs_themeish(theme)) bootstraplib::bs_global_set(old_theme)
+    if (is_bs_themeish(theme)) bslib::bs_global_set(old_theme)
   }
 
   # pre_processor
@@ -97,7 +97,7 @@ html_document_base <- function(theme = NULL,
       format_deps <- append(format_deps, list(html_dependency_jquery()))
       format_deps <- append(format_deps, bootstrap_dependencies(
         # If TRUE, a as_bs_theme(theme) has been set globally (so users may customize it)
-        if (is_bs_themeish(theme)) bootstraplib::bs_global_get() else theme
+        if (is_bs_themeish(theme)) bslib::bs_global_get() else theme
       ))
     }
     else if (isTRUE(bootstrap_compatible) && is_shiny(runtime)) {